GasFee Staleness OnChain (#412)
* Tidy up GasFeeCache contract

* Staleness

* Added revert test

* gethwrappers

* Add staleness constructor arg offchain

* Immutable staleness threshold

+++ contracts/src/v0.8/ccip/models/GE.sol
@@ -9,7 +9,7 @@ import {Internal} from "./Internal.sol";
 library GE {
   struct FeeUpdate {
     uint64 chainId;
-    uint256 linkPerUnitGas;
+    uint128 linkPerUnitGas;
   }
 
   struct ExecutionReport {
